Apply cargo fmt / clippy --fix to the BooleanFunction fix
Drops the now-redundant i128 casts clippy flagged (boolean_function_expr
already takes an i128) and reflows the touched lines.

+++ src/evaluator/function_application.rs
@@ -1443,7 +1443,7 @@ pub fn apply_curried_call(
       // function in those arguments, the same value BooleanFunction's
       // three-argument form gives.
       Ok(crate::functions::boolean_ast::boolean_function_expr(
-        *n as i128, args,
+        *n, args,
       ))
     }
     // BooleanCountingFunction[spec, n][b1, …] — True when the number of True
